refactor: replace map store with MapHelper class and change px units to rem units.
