+++ web/src/lib/components/map/Dot.svelte
